+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# eslint-overlook
+# eslint-baseline
 
 ## 0.3.0
 
 ### Minor Changes
 
-- [#3](https://github.com/lukahartwig/eslint-overlook/pull/3) [`8ba481a`](https://github.com/lukahartwig/eslint-overlook/commit/8ba481a210bd25c9bb98677d6333a68b6c525260) Thanks [@lukahartwig](https://github.com/lukahartwig)! - Added a few more eslint plugins and fixed the errors in the code.
+- [#3](https://github.com/lukahartwig/eslint-baseline/pull/3) [`8ba481a`](https://github.com/lukahartwig/eslint-baseline/commit/8ba481a210bd25c9bb98677d6333a68b6c525260) Thanks [@lukahartwig](https://github.com/lukahartwig)! - Added a few more eslint plugins and fixed the errors in the code.